fix: author_id

+++ backend/app/controllers/comments.js
@@ -6,6 +6,7 @@ exports.getAllComments = async (req, res, next) => {
         const comments = await Comment.query()
             .where('comments.post_id', req.params.id)
             .withGraphFetched('authors')
+            .orderBy('published', 'desc')
 
         res.status(200).json(comments)
     } catch (e) {
@@ -29,14 +30,18 @@ exports.getOneComment = async (req, res, next) => {
 exports.createComment = async (req, res, next) => {
     try {
         const comment = await Comment.query().insertGraph({
-            author_id: req.body.author_id,
+            author_id: req.userId,
             post_id: req.body.post_id,
             content: req.body.content,
             published: date,
             updated: date
         })
 
-        res.status(200).json("Bravo pour ton nouveau commentaire !")
+        const commentRes = await Comment.query()
+            .findById(comment.id)
+            .withGraphFetched('authors')
+
+        res.status(200).json(commentRes)
     } catch (e) {
         return next(e)
     }
@@ -44,6 +49,12 @@ exports.createComment = async (req, res, next) => {
 
 exports.updateComment = async (req, res, next) => {
     try {
+        const userComment = await Comment.query().findById(req.params.id)
+        
+        if (req.userId !== userComment.author_id && req.role === "user") {
+            return res.status(401).json({ error: '401: User is not authorizated' })
+        }
+
         const comment = await Comment.query()
             .findById(req.params.id)
             .patch({
@@ -59,6 +70,12 @@ exports.updateComment = async (req, res, next) => {
 
 exports.deleteOneComment = async (req, res, next) => {
     try {
+        const userComment = await Comment.query().findById(req.params.id)
+        
+        if (req.userId !== userComment.author_id && req.role === "user") {
+            return res.status(401).json({ error: '401: User is not authorizated' })
+        }
+
         await Comment.query().findById(req.params.id).delete()
 
         res.json("!!! Commentaire effacé !!!")
